Install files into $PREFIX/etc/bash-completion.d
The reason is that as of bash-completion 2.8 completion files stored in
$PREFIX/share/bash-completion/completions are sourced only when a user
wants to autocomplete a command, which hasn't been soured yet.
Unfortunately, it causes a lot of problems:
- The pkg completions do not overwrite the official pkg completions are
  loaded unconditionally due to being installed into
  $PREFIX/etc/bash-completion.d.
- The service completions are never sourced as the bash-completion
  library has completions for Linux-specific service command hardcoded
  inside the $PREFIX/bash-completion/bash-completion file.

The solution is to install our completions into the non-default compat
directory: $PREFIX/etc/bash-completion.d. Files in this directory are
sourced just before user-defined completions.